+++ jax/typing.py
@@ -72,4 +72,6 @@ see `Non-array inputs NumPy vs JAX`_
 from jax._src.typing import (
     ArrayLike as ArrayLike,
     DTypeLike as DTypeLike,
+    Shape as Shape,
+    DimSize as DimSize,
 )